+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewAutoRestCommandHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class AddNewAutoRestCommandHandler : AddNewCommandHandler
+    {
+    }
+}
