+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewCommandHandler.cs
@@ -1,0 +1,15 @@
+using MonoDevelop.Components.Commands;
+
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public abstract class AddNewCommandHandler : CommandHandler
+    {
+        protected override void Run()
+        {
+        }
+
+        protected override void Run(object dataItem)
+        {
+        }
+    }
+}
