--- src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewCommandHandlers.cs
@@ -1,35 +1,0 @@
-﻿using MonoDevelop.Components.Commands;
-
-namespace ApiClientCodeGen.VSMac.Commands.Handlers
-{
-    public class AddNewAutoRestCommandHandler : AddNewCommandHandler
-    {
-    }
-
-    public class AddNewNSwagCommandHandler : AddNewCommandHandler
-    {
-    }
-
-    public class AddNewNSwagStudioCommandHandler : AddNewCommandHandler
-    {
-    }
-
-    public class AddNewSwaggerCommandHandler : AddNewCommandHandler
-    {
-    }
-
-    public class AddNewOpenApiCommandHandler : AddNewCommandHandler
-    {
-    }
-
-    public abstract class AddNewCommandHandler : CommandHandler
-    {
-        protected override void Run()
-        {
-        }
-
-        protected override void Run(object dataItem)
-        {
-        }
-    }
-}
