+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewNSwagCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class AddNewNSwagCommandHandler : AddNewCommandHandler
+    {
+    }
+}
