+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewNSwagStudioCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class AddNewNSwagStudioCommandHandler : AddNewCommandHandler
+    {
+    }
+}
