+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewOpenApiCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class AddNewOpenApiCommandHandler : AddNewCommandHandler
+    {
+    }
+}
