+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/AddNewSwaggerCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class AddNewSwaggerCommandHandler : AddNewCommandHandler
+    {
+    }
+}
