+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateAutoRestCommandHandler.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+using Cairo;
+using MonoDevelop.Core;
+
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class GenerateAutoRestCommandHandler : GenerateCommandHandler
+    {
+    }
+}
