+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateNSwagCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class GenerateNSwagCommandHandler : GenerateCommandHandler
+    {
+    }
+}
