+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateOpenApiCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class GenerateOpenApiCommandHandler : GenerateCommandHandler
+    {
+    }
+}
