+++ src/ApiClientCodeGen.VSMac/Commands/Handlers/GenerateSwaggerCommandHandler.cs
@@ -1,0 +1,6 @@
+namespace ApiClientCodeGen.VSMac.Commands.Handlers
+{
+    public class GenerateSwaggerCommandHandler : GenerateCommandHandler
+    {
+    }
+}
